all: implement a more reliable image comparison
This CL introduces the internal package cmpimg which allows to reliably
compare whether two images are the same, taking into account
idiosyncracies of the gonum/plot-supported image formats.

Fixes #280
